# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Risk Control API",
        "description": "API para sistema de control de riesgo en operaciones de trading",
        "contact": {
            "email": "soporte@riskcontrol.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/incidents": {
            "get": {
                "tags": [
                    "Incidents"
                ],
                "summary": "Listar incidentes",
                "description": "Obtener lista de incidentes",
                "operationId": "c58b1625d41eb05fa2809fa6eb326e25",
                "parameters": [
                    {
                        "name": "account_id",
                        "in": "query",
                        "description": "Filtrar por ID de cuenta",
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "rule_id",
                        "in": "query",
                        "description": "Filtrar por ID de regla",
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "severity",
                        "in": "query",
                        "description": "Filtrar por severidad",
                        "schema": {
                            "type": "string",
                            "enum": [
                                "HARD",
                                "SOFT"
                            ]
                        }
                    },
                    {
                        "name": "search",
                        "in": "query",
                        "description": "Buscar por login de cuenta",
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Elementos por página",
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Número de página",
                        "schema": {
                            "type": "integer",
                            "default": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de incidentes",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/components/schemas/Incident"
                                            }
                                        },
                                        "meta": {
                                            "properties": {
                                                "total": {
                                                    "type": "integer",
                                                    "example": 50
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 20
                                                },
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "from": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "to": {
                                                    "type": "integer",
                                                    "example": 20
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
        "/accounts/{account_id}/incidents": {
            "get": {
                "tags": [
                    "Incidents"
                ],
                "summary": "Obtener incidentes por cuenta",
                "description": "Obtener incidentes por cuenta",
                "operationId": "73b6a02dbe5326a187e4639b23e973dd",
                "parameters": [
                    {
                        "name": "account_id",
                        "in": "path",
                        "description": "ID de la cuenta",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Elementos por página",
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Incidentes de la cuenta",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "account": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "login": {
                                                    "type": "integer",
                                                    "example": 21002025
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "incidents": {
                                            "type": "object",
                                            "allOf": [
                                                {
                                                    "$ref": "#/components/schemas/PaginatedResponse"
                                                },
                                                {
                                                    "properties": {
                                                        "data": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/components/schemas/Incident"
                                                            }
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
        "/rules/{rule_id}/incidents": {
            "get": {
                "tags": [
                    "Incidents"
                ],
                "summary": "Obtener incidentes por regla",
                "description": "Obtener incidentes por regla",
                "operationId": "69b57f05ad2fac68cd987d5755dcd3b6",
                "parameters": [
                    {
                        "name": "rule_id",
                        "in": "path",
                        "description": "ID de la regla",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "description": "Elementos por página",
                        "schema": {
                            "type": "integer",
                            "default": 20
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Incidentes de la regla",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "rule": {
                                            "properties": {
                                                "id": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "name": {
                                                    "type": "string",
                                                    "example": "Duración mínima"
                                                },
                                                "type": {
                                                    "type": "string",
                                                    "example": "DURATION"
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "incidents": {
                                            "type": "object",
                                            "allOf": [
                                                {
                                                    "$ref": "#/components/schemas/PaginatedResponse"
                                                },
                                                {
                                                    "properties": {
                                                        "data": {
                                                            "type": "array",
                                                            "items": {
                                                                "$ref": "#/components/schemas/Incident"
                                                            }
                                                        }
                                                    },
                                                    "type": "object"
                                                }
                                            ]
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    },
                    "404": {
                        "$ref": "#/components/responses/NotFound"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
        "/incidents/statistics": {
            "get": {
                "tags": [
                    "Incidents"
                ],
                "summary": "Estadísticas de incidentes",
                "description": "Obtener estadísticas de incidentes",
                "operationId": "7c67e491508aedf7dab3f9d5736455b4",
                "parameters": [
                    {
                        "name": "time_range",
                        "in": "query",
                        "description": "Rango de tiempo",
                        "schema": {
                            "type": "string",
                            "default": "today",
                            "enum": [
                                "today",
                                "week",
                                "month",
                                "year"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Estadísticas",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "time_range": {
                                            "type": "string",
                                            "example": "today"
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 25
                                        },
                                        "by_severity": {
                                            "properties": {
                                                "HARD": {
                                                    "type": "integer",
                                                    "example": 10
                                                },
                                                "SOFT": {
                                                    "type": "integer",
                                                    "example": 15
                                                }
                                            },
                                            "type": "object"
                                        },
                                        "top_rules": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "name": {
                                                        "type": "string",
                                                        "example": "Duración mínima"
                                                    },
                                                    "type": {
                                                        "type": "string",
                                                        "example": "DURATION"
                                                    },
                                                    "count": {
                                                        "type": "integer",
                                                        "example": 8
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "$ref": "#/components/responses/Unauthorized"
                    }
                },
                "security": [
                    {
                        "apiKey": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "PaginatedResponse": {
                "title": "Respuesta Paginada",
                "properties": {
                    "data": {
                        "type": "array",
                        "items": {
                            "type": "object"
                        }
                    },
                    "links": {
                        "type": "object"
                    },
                    "meta": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
...
            "Incident": {
                "title": "Incidencia",
                "description": "Registro de violación de una regla de riesgo",
                "required": [
                    "rule_id",
                    "account_id",
                    "severity"
                ],
                "properties": {
                    "id": {
                        "description": "Identificador único",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "rule_id": {
                        "description": "ID de la regla violada",
                        "type": "integer",
                        "example": 1
                    },
                    "account_id": {
                        "description": "ID de la cuenta involucrada",
                        "type": "integer",
                        "example": 1
                    },
                    "trade_id": {
                        "description": "ID del trade relacionado (si aplica)",
                        "type": "integer",
                        "example": 1,
                        "nullable": true
                    },
                    "severity": {
                        "description": "Severidad del incidente",
                        "type": "string",
                        "enum": [
                            "HARD",
                            "SOFT"
                        ],
                        "example": "HARD"
                    },
                    "description": {
                        "description": "Descripción del incidente",
                        "type": "string",
                        "example": "Trade cerrado en menos de 60 segundos"
                    },
                    "created_at": {
                        "description": "Fecha y hora de creación del incidente",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
...
        },
        "responses": {
...
            "Unauthorized": {
                "description": "No autorizado",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "error": {
                                    "type": "string",
                                    "example": "Unauthorized"
                                },
                                "message": {
                                    "type": "string",
                                    "example": "Clave API inválida"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            },
            "NotFound": {
                "description": "Recurso no encontrado",
                "content": {
                    "application/json": {
                        "schema": {
                            "properties": {
                                "error": {
                                    "type": "string",
                                    "example": "Not Found"
                                },
                                "message": {
                                    "type": "string",
                                    "example": "Recurso no encontrado"
                                }
                            },
                            "type": "object"
                        }
                    }
                }
            }
        },
        "securitySchemes": {
            "apiKey": {
                "type": "apiKey",
                "name": "X-API-KEY",
                "in": "header"
            }
        }
    },
...
}